% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarantino_data.R
\docType{data}
\name{tarantino_data}
\alias{tarantino_data}
\title{Profanities in Quentin Tarantino Films}
\format{
A data frame with 1895 observations on the following 4 variables.
\describe{
\item{\code{movie}}{ movie title}
\item{\code{type}}{ distinguishing words and death}
\item{\code{word}}{ curse word used (if type is word)}
\item{\code{minutes_in}}{ number of minutes into movie}
}
}
\source{
\url{https://raw.githubusercontent.com/fivethirtyeight/data/master/tarantino/tarantino.csv} 
Data collected by Andrew Flowers and downloaded via FiveThirtyEight 
"Tarantino".
}
\usage{
data(tarantino_data)
}
\description{
Data on curse words and obscenities used in various Quentin Tarantino directed films.
}
\keyword{datasets}
